More fancy config via modals

+++ src/features/guild-actions.ts
@@ -1,4 +1,4 @@
-import { ApplicationCommandType, GatewayDispatchEvents, GuildMemberFlags, InteractionType, MessageFlags, PermissionFlagsBits } from "discord-api-types/v10";
+import { ApplicationCommandType, ChannelType, ComponentType, GatewayDispatchEvents, GuildMemberFlags, InteractionType, MessageFlags, PermissionFlagsBits, SelectMenuDefaultValueType, type APIModalInteractionResponseCallbackData } from "discord-api-types/v10";
 import type { EventModule } from "../feature.ts";
 import { getSubcommandAndOptions, hasBitfield, hasBitfield2 } from "../utils.ts";
 
@@ -61,10 +61,84 @@ const guildActionsModule: EventModule = {
                 interaction.data.type === ApplicationCommandType.ChatInput &&
                 interaction.data.name === "welcome-actions"
             ) {
-                const { options } = getSubcommandAndOptions(interaction.data);
+
+                const guildActions = await db.getGuildActions(interaction.guild_id);
+                const modal: APIModalInteractionResponseCallbackData = {
+                    title: "Welcome Actions",
+                    custom_id: `welcome_actions:${interaction.guild_id}`,
+                    components: [
+                        {
+                            type: ComponentType.TextDisplay,
+                            content: "Configure the welcome actions for this server." +
+                                "\n-# - You can set a role to give to users when they complete onboarding (ie accepted rules)." +
+                                "\n-# - You can set a channel for the bot to send join/leave messages.",
+                        },
+                        {
+                            type: ComponentType.Label,
+                            label: "Role to give (optional)",
+                            description: "The role to give new members when they complete onboarding.",
+                            component: {
+                                type: ComponentType.RoleSelect,
+                                custom_id: "role",
+                                placeholder: "@member",
+                                default_values: guildActions?.join_role_id
+                                    ? [{ type: SelectMenuDefaultValueType.Role, id: guildActions.join_role_id }]
+                                    : undefined,
+                                required: false,
+                                max_values: 1,
+                                min_values: 0,
+                            },
+                        },
+                        {
+                            type: ComponentType.Label,
+                            label: "Welcome log channel (optional)",
+                            description: "The channel to send join/leave messages.",
+                            component: {
+                                type: ComponentType.ChannelSelect,
+                                custom_id: "channel",
+                                placeholder: "#moderation-logs",
+                                channel_types: [
+                                    ChannelType.GuildText,
+                                    ChannelType.GuildAnnouncement,
+                                    ChannelType.PublicThread,
+                                    ChannelType.PrivateThread,
+                                    ChannelType.AnnouncementThread,
+                                ],
+                                default_values: guildActions?.log_channel_id
+                                    ? [{ type: SelectMenuDefaultValueType.Channel, id: guildActions.log_channel_id }]
+                                    : undefined,
+                                required: false,
+                                max_values: 1,
+                                min_values: 0,
+                            },
+                        },
+                    ]
+                }
+                await api.interactions.createModal(interaction.id, interaction.token, modal);
+            } else if (
+                interaction.type === InteractionType.ModalSubmit &&
+                interaction.data.custom_id.startsWith("welcome_actions:")
+            ) {
                 const guildId = interaction.guild_id;
-                const roleId = options.role
-                const channelId = options.channel;
+                let roleId: string | null = null;
+                let channelId: string | null = null;
+                for (const component of interaction.data.components) {
+                    if (component.type === ComponentType.Label) {
+                        if (component.component.custom_id === "role") {
+                            const roleSelect = component.component;
+                            if (roleSelect.type === ComponentType.RoleSelect) {
+                                roleId = roleSelect.values?.[0] || null;
+                            }
+                        } else if (component.component.custom_id === "channel") {
+                            const channelSelect = component.component;
+                            if (channelSelect.type === ComponentType.ChannelSelect) {
+                                channelId = channelSelect.values?.[0] || null;
+                            }
+                        }
+                    }
+                }
+
+                const currentGuildActions = await db.getGuildActions(guildId);
 
                 if (!roleId && !channelId) {
                     await db.removeGuildActions(guildId);
@@ -94,7 +168,7 @@ const guildActionsModule: EventModule = {
                         return;
                     }
                 }
-                if (channelId) {
+                if (channelId && !currentGuildActions?.log_channel_id) {
                     try {
                         await api.channels.createMessage(channelId as string, {
                             content: `This channel has been set as the welcome channel for this server! I'll send join/leave messages here when users complete onboarding.`,
